cut out cloneTemplateAndCopyToProjectAsync()
